requirements.txt
##############################
# Requerimientos principales #
##############################

# Framework Web
Flask>=3.0.0

# Base de datos PostgreSQL
psycopg2-binary>=2.9.9

# Para deployment en producción
gunicorn>=21.2.0

#####################################
# Requerimientos opcionales (tests) #
#####################################
pytest>=7.4.0
pytest-cov>=4.1.0

##################################
# Utilidades recomendadas #
##################################
python-dotenv>=1.0.0
